requirements.txt
crewai
google-generativeai
python-dotenv
PyPDF2 
python-docx 
python-pptx
litellm